+++ lib/sidekiq-enqueuer.rb
@@ -1,1 +1,3 @@
-require 'sidekiq/enqueuer'
+# frozen_string_literal: true
+
+require_relative "sidekiq_enqueuer"