+++ lib/sidekiq/enqueuer/railtie.rb
@@ -1,6 +1,0 @@
-module Sidekiq::Enqueuer
-  class Railtie < ::Rails::Railtie
-    initializer "sidekiq_enqueuer_loaded" do
-    end
-  end
-end