+++ lib/sidekiq/enqueuer/web_extension/params_parser.rb
@@ -1,3 +1,7 @@
+# frozen_string_literal: true
+
+# frozen_string_literal: true
+
 module Sidekiq
   module Enqueuer
     module WebExtension
@@ -13,23 +17,28 @@ module Sidekiq
 
         def process
           worker.params.each do |expected_param|
-            value = extract_value(expected_param.name.to_s)
-            expected_param.value = value
-            raise NoProvidedValueForRequiredParam if expected_param.required? && !expected_param.value.present?
+            expected_param.value = extract_value(expected_param.name.to_s)
+
+            if expected_param.required? && !expected_param.value.present?
+              raise NoProvidedValueForRequiredParam
+            end
           end
+
           worker.params.map(&:value)
         end
 
         private
 
         def extract_value(param_name)
-          return nil unless raw_params[param_name].present?
+          return unless raw_params[param_name].present?
+
           cleanup(raw_params[param_name])
         end
 
         def cleanup(value)
-          return nil if value.to_s.downcase == 'nil'
-          return '' if value.to_s.strip.empty?
+          return if value.to_s.downcase == "nil"
+          return "" if value.to_s.strip.empty?
+
           value
         end
       end