+++ spec/support/web_extension_api_context.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+shared_context "web extension api context" do
+  include Rack::Test::Methods
+
+  subject(:app) { Sidekiq::Web }
+end